Split crypto/bake
