+++ src/CMakeLists.txt
@@ -21,7 +21,10 @@ set(src
   core/u64.c
   core/util.c
   core/word.c
-  crypto/bake.c
+  crypto/bake/bake_bmqv.c
+  crypto/bake/bake_bsts.c
+  crypto/bake/bake_bpace.c
+  crypto/bake/bake_kdf.c
   crypto/bash/bash_f.c
   crypto/bash/bash_hash.c
   crypto/bash/bash_prg.c
