+++ win/vs15/libbee2.vcxproj
@@ -373,7 +373,10 @@
     <ClCompile Include="..\..\src\core\u64.c" />
     <ClCompile Include="..\..\src\core\util.c" />
     <ClCompile Include="..\..\src\core\word.c" />
-    <ClCompile Include="..\..\src\crypto\bake.c" />
+    <ClCompile Include="..\..\src\crypto\bake\bake_bmqv.c" />
+    <ClCompile Include="..\..\src\crypto\bake\bake_bpace.c" />
+    <ClCompile Include="..\..\src\crypto\bake\bake_bsts.c" />
+    <ClCompile Include="..\..\src\crypto\bake\bake_kdf.c" />
     <ClCompile Include="..\..\src\crypto\bash\bash_prg.c" />
     <ClCompile Include="..\..\src\crypto\bash\bash_f.c" />
     <ClCompile Include="..\..\src\crypto\bash\bash_hash.c" />
